backend/requirements.txt
Django==5.2
django-ninja==1.3.0
django-unfold==0.38.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.1
PyJWT==2.8.0
google-auth==2.29.0
django-cors-headers==4.3.1
requests==2.31.0
gunicorn==21.2.0
Pillow==10.3.0
